requirements.txt
selenium
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
fastapi
uvicorn[standard]
pydantic
requests